PR change requests cont
