+++ include/pipettes/core/i2c_writer.hpp
@@ -33,7 +33,19 @@ class I2CWriter {
 
     template <typename Data>
     requires std::is_integral_v<Data>
-    void write(Data data, uint16_t device_address, uint8_t reg = 0x0) {
+    void write(Data data, uint16_t device_address) {
+        // TODO do not make register default once all of the sensor
+        // drivers have been refactored.
+        std::array<uint8_t, MAX_SIZE> max_buffer{};
+        buffering(max_buffer, data);
+        pipette_messages::WriteToI2C write_msg{.address = device_address,
+                                               .buffer = max_buffer};
+        queue->try_write(write_msg);
+    }
+
+    template <typename Data>
+    requires std::is_integral_v<Data>
+    void write(uint16_t device_address, uint8_t reg, Data data) {
         // TODO do not make register default once all of the sensor
         // drivers have been refactored.
         std::array<uint8_t, MAX_SIZE> max_buffer{};
@@ -151,6 +163,13 @@ class I2CWriter {
     void debuffering(Buffer& buffer, Data& data) {
         auto* iter = buffer.begin();
         iter = bit_utils::bytes_to_int(iter, buffer.end(), data);
+    }
+
+    template <GenericByteBuffer Buffer, typename Data>
+    requires std::is_integral_v<Data>
+    void buffering(Buffer& buffer, Data& data) {
+        auto* iter = buffer.begin();
+        iter = bit_utils::int_to_bytes(data, iter, buffer.end());
     }
 
     template <GenericByteBuffer Buffer, typename Data>
