+++ include/sensors/core/tasks/capacitive_sensor_task.hpp
@@ -32,7 +32,7 @@ class CapacitiveMessageHandler {
     }
 
     void initialize() {
-        writer.write(0x0, ADDRESS, DEVICE_ID_REGISTER);
+        writer.write(ADDRESS, DEVICE_ID_REGISTER, 0x0);
         writer.read(
             ADDRESS, [this]() { internal_callback.send_to_can(); },
             [this](auto message_a) {
@@ -42,8 +42,8 @@ class CapacitiveMessageHandler {
         // We should send a message that the sensor is in a ready state,
         // not sure if we should have a separate can message to do that
         // holding off for this PR.
-        writer.write(DEVICE_CONFIGURATION, ADDRESS, CONFIGURATION_MEASUREMENT);
-        writer.write(SAMPLE_RATE, ADDRESS, FDC_CONFIGURATION);
+        writer.write(ADDRESS, CONFIGURATION_MEASUREMENT, DEVICE_CONFIGURATION);
+        writer.write(ADDRESS, FDC_CONFIGURATION, SAMPLE_RATE);
     }
 
   private:
