+++ pipettes/tests/test_i2c_writer.cpp
@@ -31,14 +31,14 @@ SCENARIO("Test the i2c command queue writer") {
                     std::get<i2c_writer::WriteToI2C>(empty_msg);
                 auto wb_one = write_msg_four_byte.buffer;
                 REQUIRE(wb_one.size() == MAX_SIZE);
-                REQUIRE(wb_one[4] == four_byte_data);
+                REQUIRE(wb_one[3] == four_byte_data);
 
                 queue.try_read(&empty_msg);
                 auto write_msg_two_byte =
                     std::get<i2c_writer::WriteToI2C>(empty_msg);
                 auto wb_two = write_msg_two_byte.buffer;
                 REQUIRE(wb_two.size() == MAX_SIZE);
-                REQUIRE(wb_two[2] == two_byte_data);
+                REQUIRE(wb_two[1] == two_byte_data);
             }
         }
         WHEN("we read messages") {
